bugfix(app):Adds refactor to code base
