+++ lib/screens/app/index.dart
@@ -27,15 +27,10 @@ class AppIndexScreen extends StatefulWidget {
 class _AppIndexScreenState extends State<AppIndexScreen> {
   int selectedScreen = 0;
 
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  final TextEditingController _nameController = TextEditingController();
-  final TextEditingController _costController = TextEditingController();
-  final TextEditingController _dateController = TextEditingController();
-
   final ExpenseAppService expenseService = ExpenseAppService();
 
   List<Widget> get _buildBody => <Widget>[
-    DashboardScreen(transactions: recentTransactions),
+    DashboardScreen(),
     StatisticScreen(),
     WalletScreen(),
     ProfileScreen(),
@@ -48,25 +43,6 @@ class _AppIndexScreenState extends State<AppIndexScreen> {
   }
 
 
-  List<dynamic> itemListApi = <dynamic>[];
-  String? varsity;
-  double grandTotal = 0.0;
-
-  double currentMonth = 0.0;
-  double lastMonth = 0.0;
-  double? monthlyChange;
-  List<Map<String, dynamic>> recentTransactions = <Map<String, dynamic>>[];
-
-
-  Future<void> loadJson() async {
-    final String response = await rootBundle.loadString('assets/data/expense_categories.json');
-    final dynamic data = json.decode(response);
-    setState(() {
-      itemListApi = data;
-    });
-  }
-
-  /*
   void _clearAll() async {
     await expenseService.truncateExpenses();
   }
@@ -74,79 +50,21 @@ class _AppIndexScreenState extends State<AppIndexScreen> {
   void _deleteTable() async {
     await expenseService.dropExpensesTable();
   }
-  */
-
-  Future<void> _getDbTotal() async {
-    final List<Map<String, dynamic>> result = await expenseService.getAllExpensesFromDB();
-
-    final double total = result.fold<double>(
-      0.0,
-          (double sum, Map<String, dynamic> e) => sum + (e['cost'] as num).toDouble(),
-    );
-
-    setState(() {
-      grandTotal = total;
-    });
-  }
-
-  void _clearTextField() {
-    _nameController.clear();
-    _costController.clear();
-    _dateController.clear();
-  }
-
-  Future<void> _loadRecentTransactions() async {
-    final List<Map<String, dynamic>> recent = await expenseService.getLast20Expenses();
-
-    setState(() {
-      recentTransactions = recent;
-    });
-  }
-
-  String _getCategoryName(String? categoryId) {
-    final dynamic category = itemListApi.firstWhere(
-          (e) => e['id'] == categoryId,
-      orElse: () => <String, String>{'description': 'Unknown'},
-    );
-    return category['description'];
-  }
-
-  void _saveExpense(Map<String, dynamic> data) async {
-    await expenseService.insertExpense(
-      name: data['name'],
-      date: data['date'],
-      categoryId: data['category'],
-      cost: double.tryParse(data['cost']) ?? 0.0,
-      categoryName: _getCategoryName(data['category']),
-    );
-
-    _clearTextField();
-    await _getDbTotal(); // Update grand total
-    await _loadRecentTransactions(); // Update transaction list
-  }
-
 
   @override
   void initState() {
     super.initState();
-    loadJson();
     //_clearAll();
     //_deleteTable();
-    _getDbTotal();
-    _loadRecentTransactions();
   }
 
   @override
   void dispose() {
     super.dispose();
-    _nameController.dispose();
-    _costController.dispose();
-    _dateController.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    final bool isLimitReached = grandTotal > 100000;
 
     return Scaffold(
       backgroundColor: Colors.white,
@@ -193,79 +111,8 @@ class _AppIndexScreenState extends State<AppIndexScreen> {
           )
         ],
       ),
-      floatingActionButton: FloatingActionButton(
-          onPressed: () async {
-
-            if (isLimitReached) {
-              _showMyDialog();
-              return;
-            }
-
-            final result = await showModalBottomSheet(
-                isDismissible: false,
-                isScrollControlled: true,
-                backgroundColor: Colors.transparent,
-                context: context,
-                builder: (BuildContext context) {
-                  return ExpenseBottomSheet(
-                    formKey: _formKey,
-                    itemNameController: _nameController,
-                    itemCostController: _costController,
-                    itemDateController: _dateController,
-                    itemListApi: itemListApi,
-                    varsity: varsity,
-                    onCategoryChanged: (String? val) {
-                      setState(() => varsity = val);
-                    },
-                  );
-                }
-            );
-            if (result != null) {
-              _saveExpense(result);
-            }
-          },
-          elevation: 0, highlightElevation: 0,
-          backgroundColor: Colors.transparent,
-          child: Container(
-            width: 64, height: 64,
-            decoration: BoxDecoration(
-              color: Colors.green,
-              borderRadius: BorderRadius.circular(64),
-            ),
-            child: Center(child: Icon(Icons.add, color: Colors.white)),
-          )
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
     );
   }
 
-  Future<void> _showMyDialog() async {
-    return showDialog<void>(
-      context: context,
-      barrierDismissible: false,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Eish Ja! neh...'),
-          content: const SingleChildScrollView(
-            child: ListBody(
-              children: <Widget>[
-                Text('Looks like you need a premium account.'),
-                SizedBox(height: 10),
-                Text('Askies neh! Contact aceMedia'),
-              ],
-            ),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Ok'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
 
 }
