+++ pubspec.yaml
@@ -36,12 +36,17 @@ dependencies:
   cupertino_icons: ^1.0.8
   path: ^1.9.1
   intl: ^0.20.2
+  lottie: ^3.3.1
   sqflite: ^2.4.2
   fl_chart: ^1.0.0
+  timezone: ^0.10.1
   flutter_svg: ^2.2.0
+  flutter_timezone: ^4.1.1
   device_info_plus: ^11.5.0
+  shared_preferences: ^2.5.3
   cached_network_image: ^3.4.1
   smooth_page_indicator: ^1.2.1
+  flutter_local_notifications: ^18.0.1
 
 dev_dependencies:
   flutter_test:
